tests for LayerNorm
