+++ week2/pos.py
@@ -1,0 +1,17 @@
+"""
+visualize:
+"""
+
+
+
+
+
+
+
+
+
+
+
+
+
+#  output:  pos.png
