+++ week2/tests/test_10.py
@@ -1,3 +1,3 @@
 """
-test for GPTv7
+running question: why do we need Dropout?
 """
