+++ week2/tests/test_7.py
@@ -1,11 +1,24 @@
 """
-test for: GPTv4
+Running question: why do we need more than one head?
 """
 
 
-def test_pos_embedding_table_is_learned():
-    """
-    :return:
-    """
+# test: different heads learn different things
+def test_different_heads_learn_different_things():
+    pass
+
+
+# test: multi-head ver 2 is logically the same as multi-head ver 1
+def test_multi_head_ver_2_is_logically_the_same_as_multi_head_ver_1():
+    pass
+
+
+# test: multi-head ver 2 is faster than ver 1
+def test_multi_head_ver_2_is_faster_than_ver_1():
+    pass
+
+
+# test: gpt learns faster with multi-head
+def test_gpt_v4_learns_faster_with_multi_head():
     pass
