+++ week2/tests/test_8.py
@@ -1,3 +1,18 @@
 """
-test for GPTv5
+Running question: why do we need Feedforward?
 """
+
+
+# test: head alone is linear
+def test_head_alone_is_linear():
+    pass
+
+
+# test: Feedforward is non-linear
+def test_feedforward_is_non_linear():
+    pass
+
+
+# test:
+def test_gpt_v4_learns_faster_with_feedforward():
+    pass
